+++ src/main/java/com/tencentcloudapi/iai/v20180301/models/CompareFaceRequest.java
@@ -85,7 +85,7 @@ B 图片的 Url、Image必须提供一个，如果都提供，只使用 Url。
     */
     @SerializedName("QualityControl")
     @Expose
-    private Integer QualityControl;
+    private Long QualityControl;
 
     /**
      * 获取A 图片 base64 数据，base64 编码后大小不可超过5M。
@@ -249,7 +249,7 @@ B 图片的 Url、Image必须提供一个，如果都提供，只使用 Url。
 默认 0。 
 若图片质量不满足要求，则返回结果中会提示图片质量检测不符要求。
      */
-    public Integer getQualityControl() {
+    public Long getQualityControl() {
         return this.QualityControl;
     }
 
@@ -271,7 +271,7 @@ B 图片的 Url、Image必须提供一个，如果都提供，只使用 Url。
 默认 0。 
 若图片质量不满足要求，则返回结果中会提示图片质量检测不符要求。
      */
-    public void setQualityControl(Integer QualityControl) {
+    public void setQualityControl(Long QualityControl) {
         this.QualityControl = QualityControl;
     }
